+++ private/Get-AzDFV2Credential.ps1
@@ -5,6 +5,13 @@ function Get-AzDFV2Credential {
     )
     Write-Debug "BEGIN: Get-AzDFV2Credential"
 
+    # Retrieve all credentials via API without parsing
+    $token = Get-AzAccessToken -ResourceUrl 'https://management.azure.com'
+    $tokenStr = [System.Net.NetworkCredential]::new('', $token.Token).Password
+    $authHeader = @{
+        'Content-Type'  = 'application/json'
+        'Authorization' = 'Bearer ' + $tokenStr
+    }
     $url = "$($script:BaseApiUrl)$($adfi.DataFactoryId)/credentials?api-version=2018-06-01"
 
     # Retrieve all credentials via Rest API
@@ -14,7 +21,7 @@ function Get-AzDFV2Credential {
         return $null
     }
 
-    [System.Collections.ArrayList] $items = @{}
+    [System.Collections.ArrayList] $items = @()
     ($r.Content | ConvertFrom-Json).value | ForEach-Object { $i = [AdfPSCredential]::New($_); $items.Add($i) | Out-Null; }
 
     Write-Debug "END: Get-AzDFV2Credential()"