+++ private/Remove-AdfObjectRestAPI.ps1
@@ -8,6 +8,12 @@ function Remove-AdfObjectRestAPI {
 
     Write-Debug "BEGIN: Remove-AdfObjectRestAPI()"
 
+    $token = Get-AzAccessToken -ResourceUrl 'https://management.azure.com'
+    $tokenStr = [System.Net.NetworkCredential]::new('', $token.Token).Password
+    $authHeader = @{
+        'Content-Type'  = 'application/json'
+        'Authorization' = 'Bearer ' + $tokenStr
+    }
     $url = "$($script:BaseApiUrl)$($adfInstance.Id)/$type_plural/$($name)?api-version=2018-06-01"
 
     # Delete given object via Rest API