+++ test/!RunAllTests.ps1
@@ -8,6 +8,9 @@ Param(
     [Parameter(Mandatory=$false)]
     [Switch]$MajorRelease,
 
+    [Parameter(Mandatory=$false)]
+    [string]$ResultOutputFormat = "NUnitXml",
+
     [Parameter(Mandatory=$true)]
     [Switch]$InstallModules
 )
@@ -70,7 +73,7 @@ try {
     $configuration.Run.Exit = $true
     $configuration.Should.ErrorAction = 'Continue'
     $configuration.CodeCoverage.Enabled = $false
-    $configuration.TestResult.OutputFormat = "NUnitXml"
+    $configuration.TestResult.OutputFormat = "JUnitXml"
     $configuration.TestResult.OutputPath = "$folder\TEST-Results.xml"
     $configuration.TestResult.Enabled = $true
     $configuration.Output.Verbosity = 'Detailed'